requirements.txt
fastapi==0.99.1
uvicorn==0.23.2
pytest==7.4.2
httpx==0.24.1
python-multipart==0.0.6
pypdf==3.15.2
reportlab>=3.6.12
google-api-python-client>=2.99.0
google-auth>=2.21.0
google-auth-oauthlib>=1.1.0
google-cloud-discoveryengine>=0.11.0
